Phil parameter controls crystal size.
New parameter crystal.length_um specifies the path length in microns
for the beam passing through the crystal.  This is intended as a quick
control for the overall scale of Bragg spot intensities.  Default is 4 microns, same
length as in NESAP benchmark.  Background (water/air) is not affected
as it is controlled separately.