Check each node is responsive at start of each script
Very basic node checking consists of asking each requested rank to
report which rank on which node it is. The report is given a time limit
of 2 minutes. If any node does not produce the expected number of
responses by this time limit, we record the node name in the
nodelist_exclude file and add it to a flag --exclude=<nodeids> to be
incorporated into the srun command. Every srun step in this job will
then exclude the unresponsive node.

Note, so far nothing has tripped this test, but it's very short to run
if it succeeds so I'm keeping it in anyway.