Change the way peak cutoff is hard coded.
Before, the anomalous peaks for Ca and Mn were cutoff below 7 and 18
sigmas; regardless of the strength of the data.  In this revision the
cutoff values are expressed as a fraction of the maximum peak height in
the map, namely Ca=0.3 and Mn=0.7.  This might still be a questionable
choice but better than before.